# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_controller.AnyMap:
    additionalProperties: true
    type: object
...
  internal_controller.runInspectionReq:
    properties:
      options:
        $ref: '#/definitions/k8s-platform-backend_internal_service.RunInspectionOptions'
      server_ids:
        items:
          type: integer
        minItems: 1
        type: array
      target_snapshot:
        items:
          $ref: '#/definitions/k8s-platform-backend_internal_service.InspectionTargetRef'
        type: array
      template_id:
        type: integer
    required:
    - server_ids
    - template_id
    type: object
...
  internal_controller.saveScheduleReq:
    properties:
      cron:
        type: string
      id:
        type: integer
      name:
        type: string
      server_ids:
        items:
          type: integer
        type: array
      status:
        type: string
      template_id:
        type: integer
    required:
    - cron
    - name
    - template_id
    type: object
  internal_controller.saveTemplateReq:
    properties:
      category:
        type: string
      checks:
        items:
          $ref: '#/definitions/k8s-platform-backend_internal_service.InspectionCheck'
        type: array
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      recommended:
        type: boolean
      tags:
        items:
          type: string
        type: array
      version:
        type: string
    required:
    - name
    type: object
...
  k8s-platform-backend_internal_service.InspectionCheck:
    properties:
      description:
        type: string
      enabled:
        type: boolean
      expect:
        additionalProperties: true
        type: object
      id:
        type: integer
      name:
        type: string
      script:
        type: string
      severity:
        description: low|medium|high|critical
        type: string
      tags:
        items:
          type: string
        type: array
      timeout_sec:
        type: integer
      type:
        description: shell|port_check|http_check
        type: string
    type: object
  k8s-platform-backend_internal_service.InspectionTargetRef:
    properties:
      id:
        type: integer
      ip:
        type: string
      name:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
...
  k8s-platform-backend_internal_service.RunInspectionOptions:
    properties:
      concurrency:
        type: integer
      failure_policy:
        type: string
      generate_report:
        type: boolean
      timeout_sec:
        type: integer
    type: object
...
  k8s-platform-backend_pkg_resp.Result:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 星枢K8S管理平台后端接口文档（统一响应结构：code/message/data）
  title: 星枢K8S管理平台 API
  version: "1.0"
paths:
...
  /automation/inspection-reports:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 巡检报告列表
      tags:
      - 巡检
  /automation/inspection-reports/{id}:
    get:
      parameters:
      - description: 报告ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 巡检报告详情
      tags:
      - 巡检
  /automation/inspection-schedules:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 巡检计划列表
      tags:
      - 巡检
    post:
      consumes:
      - application/json
      parameters:
      - description: 计划数据
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_controller.saveScheduleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 保存巡检计划
      tags:
      - 巡检
  /automation/inspection-templates:
    get:
      parameters:
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 分类
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 巡检模板列表
      tags:
      - 巡检
    post:
      consumes:
      - application/json
      parameters:
      - description: 模板数据
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_controller.saveTemplateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 保存巡检模板
      tags:
      - 巡检
  /automation/inspection-templates/{id}:
    delete:
      parameters:
      - description: 模板ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 删除巡检模板
      tags:
      - 巡检
    get:
      parameters:
      - description: 模板ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 巡检模板详情
      tags:
      - 巡检
  /automation/inspections/overview:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 巡检中心概览
      tags:
      - 巡检
  /automation/inspections/run:
    post:
      consumes:
      - application/json
      parameters:
      - description: 执行参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_controller.runInspectionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 立即执行巡检
      tags:
      - 巡检
...
  /dashboard/clusters/{id}/certificate-risks:
    get:
      consumes:
      - application/json
      description: 根据集群 ID 获取集群证书风险信息，不阻塞概览首屏加载
      parameters:
      - description: 集群ID
        example: 1001
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_controller.AnyMap'
                  type: array
              type: object
        default:
          description: 失败响应
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 集群证书风险
      tags:
      - 仪表盘接口
  /dashboard/clusters/{id}/overview:
    get:
      consumes:
      - application/json
      description: 根据集群 ID 获取集群概览统计信息（节点/Pod/工作负载/资源使用率等）
      parameters:
      - description: 集群ID
        example: 1001
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/internal_controller.AnyMap'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 集群概览
      tags:
      - 仪表盘接口
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"